move some shared functions into the utils module
